+++ src/deployer/MinimalContract.ts
@@ -1,5 +1,7 @@
 import { ethers, InterfaceAbi, Overrides, Signer } from "ethers";
 
+import { HardhatRuntimeEnvironment } from "hardhat/types";
+
 import { isFullyQualifiedName } from "hardhat/utils/contract-names";
 
 import { Linker } from "./Linker";
@@ -9,12 +11,11 @@ import { catchError, fillParameters, getChainId, getInterfaceOnlyWithConstructor
 import { MigrateError } from "../errors";
 
 import { MigrationMetadata } from "../types/tools";
-import { MigrateConfig } from "../types/migrations";
 import { ContractDeployTxWithName, OverridesAndLibs } from "../types/deployer";
 
 import { Stats } from "../tools/Stats";
-import { reporter } from "../tools/reporters/Reporter";
-import { transactionRunner } from "../tools/runners/TransactionRunner";
+import { Reporter } from "../tools/reporters/Reporter";
+import { TransactionRunner } from "../tools/runners/TransactionRunner";
 import { ArtifactProcessor } from "../tools/storage/ArtifactProcessor";
 import { TransactionProcessor } from "../tools/storage/TransactionProcessor";
 import { VerificationProcessor } from "../tools/storage/VerificationProcessor";
@@ -25,7 +26,7 @@ export class MinimalContract {
   private readonly _interface;
 
   constructor(
-    private readonly _config: MigrateConfig,
+    private readonly _hre: HardhatRuntimeEnvironment,
     private _bytecode: string,
     private readonly _abi: InterfaceAbi,
     private readonly _contractName: string = "",
@@ -49,7 +50,7 @@ export class MinimalContract {
 
     const tx = await this._createDeployTransaction(args, parameters);
 
-    if (this._config.continue) {
+    if (this._hre.config.migrate.continue) {
       return this._recoverContractAddress(tx, args);
     } else {
       return this._processContractDeploymentTransaction(tx, args);
@@ -58,11 +59,11 @@ export class MinimalContract {
 
   private async _tryLinkLibraries(parameters: OverridesAndLibs): Promise<void> {
     try {
-      if (Linker.isBytecodeNeedsLinking(this._bytecode)) {
+      if (Linker?.isBytecodeNeedsLinking(this._bytecode)) {
         return;
       }
 
-      this._bytecode = await Linker.tryLinkBytecode(this._contractName, this._bytecode, parameters.libraries || {});
+      this._bytecode = (await Linker?.tryLinkBytecode(this._contractName, this._bytecode, parameters.libraries || {}))!;
     } catch (e: any) {
       throw new MigrateError(
         `Unable to link libraries for ${this._contractName}! Try manually deploy the libraries and link them.\n Error: ${e.message}`,
@@ -83,18 +84,18 @@ export class MinimalContract {
 
   private async _recoverContractAddress(tx: ContractDeployTxWithName, args: any[]): Promise<string> {
     try {
-      const contractAddress = await TransactionProcessor.tryRestoreContractAddressByKeyFields(tx);
+      const contractAddress = await TransactionProcessor?.tryRestoreContractAddressByKeyFields(tx);
 
-      reporter!.notifyContractRecovery(tx.contractName, contractAddress);
+      Reporter!.notifyContractRecovery(tx.contractName, contractAddress!);
 
-      await this._saveContractForVerification(contractAddress, tx, args);
+      await this._saveContractForVerification(contractAddress!, tx, args);
 
-      return contractAddress;
+      return contractAddress!;
     } catch {
       /* empty */
     }
 
-    reporter!.notifyDeploymentInsteadOfRecovery(tx.contractName);
+    Reporter!.notifyDeploymentInsteadOfRecovery(tx.contractName);
 
     return this._processContractDeploymentTransaction(tx, args);
   }
@@ -104,7 +105,7 @@ export class MinimalContract {
 
     const txResponse = await signer.sendTransaction(tx);
 
-    await transactionRunner!.reportTransactionResponse(txResponse, tx.contractName);
+    await TransactionRunner!.reportTransactionResponse(txResponse, tx.contractName);
 
     const contractAddress = (await txResponse.wait(0))!.contractAddress;
 
@@ -117,9 +118,10 @@ export class MinimalContract {
     const saveMetadata: MigrationMetadata = {
       migrationNumber: Stats.currentMigration,
       contractName: tx.contractName,
+      fullyQualifiedContractName: this._getFullyQualifiedName(tx) || undefined,
     };
 
-    TransactionProcessor.saveDeploymentTransaction(tx, tx.contractName, contractAddress, saveMetadata);
+    TransactionProcessor?.saveDeploymentTransaction(tx, tx.contractName, contractAddress, saveMetadata);
 
     return contractAddress;
   }
@@ -129,21 +131,33 @@ export class MinimalContract {
       return;
     }
 
-    try {
-      let contractName = tx.contractName;
+    const contractName = this._getFullyQualifiedName(tx);
 
-      if (!isFullyQualifiedName(contractName)) {
-        contractName = ArtifactProcessor.tryGetContractName(this._rawBytecode);
+    if (contractName === null) {
+      Reporter!.reportVerificationFailedToSave(tx.contractName);
+
+      return;
+    }
+
+    VerificationProcessor.saveVerificationFunction({
+      contractAddress,
+      contractName: contractName,
+      constructorArguments: args,
+      chainId: Number(await getChainId()),
+    });
+
+    await ArtifactProcessor.saveArtifactIfNotExist(this._hre, contractName, this._rawBytecode);
+  }
+
+  private _getFullyQualifiedName(tx: ContractDeployTxWithName): string | null {
+    try {
+      if (!isFullyQualifiedName(tx.contractName)) {
+        return ArtifactProcessor.tryGetContractName(this._rawBytecode);
       }
 
-      VerificationProcessor.saveVerificationFunction({
-        contractAddress,
-        contractName: contractName,
-        constructorArguments: args,
-        chainId: Number(await getChainId()),
-      });
+      return tx.contractName;
     } catch {
-      reporter!.reportVerificationFailedToSave(tx.contractName);
+      return null;
     }
   }
 }