+++ src/deployer/adapters/EthersFactoryAdapter.ts
@@ -2,6 +2,8 @@ import { Addressable, ContractFactory, Interface } from "ethers";
 
 import { AbstractEthersAdapter } from "./AbstractEthersAdapter";
 
+import { OverridesAndName } from "../../types/deployer";
+
 import { catchError } from "../../utils";
 
 import { UNKNOWN_CONTRACT_NAME } from "../../constants";
@@ -18,10 +20,18 @@ export class EthersFactoryAdapter extends AbstractEthersAdapter {
     return JSON.stringify(instance.interface.fragments);
   }
 
-  public getContractName(instance: ContractFactory): string {
+  public getContractName(instance: ContractFactory, parameters: OverridesAndName): string {
+    if (parameters.name) {
+      return parameters.name;
+    }
+
     try {
       return ArtifactProcessor.tryGetContractName(this.getRawBytecode(instance));
     } catch {
+      if ((instance as any).contractName) {
+        return (instance as any).contractName;
+      }
+
       return UNKNOWN_CONTRACT_NAME;
     }
   }