+++ src/index.ts
@@ -1,9 +1,8 @@
 import "@nomicfoundation/hardhat-verify";
 
 import { ActionType } from "hardhat/types";
-import { lazyObject } from "hardhat/plugins";
 import { TASK_CLEAN, TASK_COMPILE } from "hardhat/builtin-tasks/task-names";
-import { extendConfig, extendEnvironment, task, types } from "hardhat/config";
+import { extendConfig, task, types } from "hardhat/config";
 
 import "./type-extensions";
 
@@ -12,14 +11,14 @@ import { TASK_MIGRATE, TASK_MIGRATE_VERIFY } from "./constants";
 
 import { MigrateConfig, MigrateVerifyConfig } from "./types/migrations";
 
-import { DefaultStorage, MigrateStorage } from "./tools/storage/MigrateStorage";
+import { DefaultStorage } from "./tools/storage/MigrateStorage";
 import { VerificationProcessor } from "./tools/storage/VerificationProcessor";
 
 import { Migrator } from "./migrator/Migrator";
 import { Verifier } from "./verifier/Verifier";
 
 export { Deployer } from "./deployer/Deployer";
-export { DefaultStorage } from "./tools/storage/MigrateStorage";
+export { UserStorage } from "./tools/storage/MigrateStorage";
 export { PublicReporter as Reporter } from "./tools/reporters/PublicReporter";
 
 extendConfig(migrateConfigExtender);
@@ -33,7 +32,7 @@ const migrate: ActionType<MigrateConfig> = async (taskArgs, env) => {
     force: env.config.migrate.force,
   });
 
-  await Migrator.initialize(env);
+  await Migrator.buildMigrateTaskDeps(env);
 
   await new Migrator(env).migrate();
 
@@ -45,23 +44,15 @@ const migrate: ActionType<MigrateConfig> = async (taskArgs, env) => {
 const migrateVerify: ActionType<MigrateVerifyConfig> = async (taskArgs, env) => {
   const config = extendVerifyConfigs(taskArgs);
 
-  await Verifier.initialize(env);
+  await Verifier.buildVerifierTaskDeps(env);
 
   await new Verifier(env, config).verifyBatch(
     VerificationProcessor.restoreSavedVerificationFunctions(config.inputFile),
   );
 };
 
-extendEnvironment((hre) => {
-  hre.migrator = lazyObject(() => {
-    return new Migrator(hre);
-  });
-
-  hre.storage = lazyObject(() => DefaultStorage);
-});
-
 task(TASK_CLEAN, "Clears the cache and deletes all artifacts").setAction(async (conf, hre, runSuper) => {
-  MigrateStorage.clean();
+  DefaultStorage.clean();
 
   await runSuper();
 });