+++ src/migrator/Migrator.ts
@@ -12,24 +12,23 @@ import { MigrateError } from "../errors";
 
 import { MigrateConfig } from "../types/migrations";
 
-import { Linker } from "../deployer/Linker";
 import { Deployer } from "../deployer/Deployer";
-import { Verifier } from "../verifier/Verifier";
+
+import { createLinker } from "../deployer/Linker";
 
 import { Stats } from "../tools/Stats";
 
-import { initReporter, reporter } from "../tools/reporters/Reporter";
-import { transactionRunner } from "../tools/runners/TransactionRunner";
+import { TransactionRunner } from "../tools/runners/TransactionRunner";
+import { createAndInitReporter, Reporter } from "../tools/reporters/Reporter";
 
-import { initNetworkManager } from "../tools/network/NetworkManager";
+import { buildNetworkDeps } from "../tools/network/NetworkManager";
 
-import { TransactionProcessor } from "../tools/storage/TransactionProcessor";
-import { MigrateStorage } from "../tools/storage/MigrateStorage";
+import { clearAllStorage } from "../tools/storage/MigrateStorage";
 import { ArtifactProcessor } from "../tools/storage/ArtifactProcessor";
+import { createTransactionProcessor } from "../tools/storage/TransactionProcessor";
 
 export class Migrator {
   private readonly _deployer: Deployer;
-  private readonly _verifier: Verifier;
 
   private readonly _migrationFiles: string[];
 
@@ -38,27 +37,23 @@ export class Migrator {
     private _config: MigrateConfig = _hre.config.migrate,
   ) {
     this._deployer = new Deployer(_hre);
-    this._verifier = new Verifier(_hre, {
-      parallel: this._config.verifyParallel,
-      attempts: this._config.verifyAttempts,
-    });
 
     this._migrationFiles = this._getMigrationFiles();
   }
 
   public async migrate() {
-    reporter!.reportMigrationBegin(this._migrationFiles);
+    Reporter!.reportMigrationBegin(this._migrationFiles);
 
     for (const element of this._migrationFiles) {
       Stats.currentMigration = this._getMigrationNumber(element);
 
-      reporter!.reportMigrationFileBegin(element);
+      Reporter!.reportMigrationFileBegin(element);
 
       try {
         // eslint-disable-next-line @typescript-eslint/no-var-requires
         const migration = require(resolvePathToFile(this._config.pathToMigrations, element));
 
-        await migration(this._deployer, this._verifier);
+        await migration(this._deployer);
       } catch (e: unknown) {
         if (e instanceof MigrateError) {
           throw new HardhatPluginError(pluginName, e.message, e);
@@ -68,7 +63,7 @@ export class Migrator {
       }
     }
 
-    transactionRunner!.summary();
+    TransactionRunner!.summary();
   }
 
   private _getMigrationFiles() {
@@ -107,15 +102,15 @@ export class Migrator {
     return parseInt(basename(file));
   }
 
-  public static async initialize(hre: HardhatRuntimeEnvironment): Promise<void> {
-    Linker.setConfig(hre.config.migrate);
-    TransactionProcessor.setConfig(hre.config.migrate);
+  public static async buildMigrateTaskDeps(hre: HardhatRuntimeEnvironment): Promise<void> {
+    createLinker(hre);
+    createTransactionProcessor(hre.config.migrate);
 
-    initNetworkManager(hre);
-    await initReporter(hre.config.migrate);
+    buildNetworkDeps(hre);
+    await createAndInitReporter(hre);
 
     if (!hre.config.migrate.continue) {
-      MigrateStorage.clearAll();
+      clearAllStorage();
     }
 
     await ArtifactProcessor.parseArtifacts(hre);