+++ src/tools/network/NetworkManager.ts
@@ -4,16 +4,17 @@ import type { HardhatEthersProvider as HardhatEthersProviderT } from "@nomicfoun
 
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 
-import { ethersProvider, initEthersProvider } from "./EthersProvider";
+import { ethersProvider, createEthersProvider } from "./EthersProvider";
 
-import { reporter } from "../reporters/Reporter";
+import { Reporter } from "../reporters/Reporter";
+import { createTransactionRunner } from "../runners/TransactionRunner";
 
 import { sleep } from "../../utils";
 import { MAX_RECONNECT_ATTEMPTS, RECONNECT_INTERVAL } from "../../constants";
 
-import { initTransactionRunner } from "../runners/TransactionRunner";
-
 class StateMiddleware {
+  private static _isNetworkIssue: boolean = false;
+
   public static async retry<T extends (...args: any[]) => any>(
     fn: T,
     args: Parameters<T>,
@@ -22,20 +23,28 @@ class StateMiddleware {
     try {
       const result = await fn(...args);
 
-      reporter?.resetSpinnerMessageIfActive();
+      if (this._isNetworkIssue) {
+        Reporter?.stopSpinner();
+
+        this._isNetworkIssue = false;
+      }
 
       return result;
     } catch (e: any) {
-      // TODO: use spinner instead of console.log.
       const networkErrorCodes = ["EAI_AGAIN", "ENETDOWN", "ENETUNREACH", "ENOTFOUND", "ECONNABORTED"];
       const isNetworkError = networkErrorCodes.includes(e.code) || e.isAxiosError;
 
       if (!isNetworkError) {
+        Reporter?.stopSpinner();
+
         throw e;
       }
 
-      // TODO: set timeout manually.
-      reporter!.reportNetworkError(retryCount, fn.name, e);
+      await Reporter?.startSpinner("network-error");
+
+      Reporter!.reportNetworkError(retryCount, fn.name, e);
+
+      this._isNetworkIssue = true;
 
       await sleep(RECONNECT_INTERVAL);
 
@@ -46,8 +55,18 @@ class StateMiddleware {
       return this.retry(fn, args, retryCount + 1);
     }
   }
+}
 
-  public static withRetry<T extends { [key: string]: any }>(instance: T): T {
+class NetworkManager {
+  public axios: Axios;
+  public provider: HardhatEthersProviderT;
+
+  constructor() {
+    this.axios = this.withRetry(axios);
+    this.provider = this.withRetry(ethersProvider!);
+  }
+
+  public withRetry<T extends { [key: string]: any }>(instance: T): T {
     return new Proxy(instance, {
       get(target, propKey, receiver) {
         const origMethod = target[propKey as keyof T];
@@ -64,21 +83,11 @@ class StateMiddleware {
   }
 }
 
-class NetworkManager {
-  public axios: Axios;
-  public provider: HardhatEthersProviderT;
-
-  constructor() {
-    this.axios = StateMiddleware.withRetry(axios);
-    this.provider = StateMiddleware.withRetry(ethersProvider!);
-  }
-}
-
 export let networkManager: NetworkManager | null = null;
 
-export function initNetworkManager(hre: HardhatRuntimeEnvironment) {
-  initEthersProvider(hre);
-  initTransactionRunner(hre);
+export function buildNetworkDeps(hre: HardhatRuntimeEnvironment) {
+  createEthersProvider(hre);
+  createTransactionRunner(hre);
 
   if (networkManager) {
     return;
@@ -87,6 +96,9 @@ export function initNetworkManager(hre: HardhatRuntimeEnvironment) {
   networkManager = new NetworkManager();
 }
 
+/**
+ * Used only in test environments to ensure test atomicity
+ */
 export function resetNetworkManager() {
   networkManager = null;
 }