+++ src/tools/storage/ArtifactProcessor.ts
@@ -1,5 +1,7 @@
 import { Artifact, HardhatRuntimeEnvironment } from "hardhat/types";
 
+import { isFullyQualifiedName } from "hardhat/utils/contract-names";
+
 import { ArtifactStorage } from "./MigrateStorage";
 
 import { MigrateError } from "../../errors";
@@ -9,8 +11,8 @@ import { bytecodeHash, catchError } from "../../utils";
 import { ArtifactExtended, NeededLibrary } from "../../types/deployer";
 
 @catchError
-export class ArtifactProcessor {
-  public static async parseArtifacts(_hre: HardhatRuntimeEnvironment): Promise<void> {
+class BaseArtifactProcessor {
+  public async parseArtifacts(_hre: HardhatRuntimeEnvironment): Promise<void> {
     ArtifactStorage.clear();
 
     const names = await _hre.artifacts.getAllFullyQualifiedNames();
@@ -32,7 +34,23 @@ export class ArtifactProcessor {
     }
   }
 
-  public static tryGetArtifactByName(contractName: string): ArtifactExtended {
+  public async saveArtifactIfNotExist(
+    _hre: HardhatRuntimeEnvironment,
+    contractName: string,
+    bytecode?: string,
+  ): Promise<void> {
+    if (!isFullyQualifiedName(contractName) || (bytecode ? true : ArtifactStorage.get(bytecodeHash(bytecode!)))) {
+      return;
+    }
+
+    const artifact = await _hre.artifacts.readArtifact(contractName);
+
+    const contract: ArtifactExtended = { ...artifact, neededLibraries: this._parseLibrariesOfArtifact(artifact) };
+
+    ArtifactStorage.set(bytecode ? bytecodeHash(bytecode) : bytecodeHash(artifact.bytecode), contract, true);
+  }
+
+  public tryGetArtifactByName(contractName: string): ArtifactExtended {
     const artifact = ArtifactStorage.get(contractName);
 
     if (!artifact) {
@@ -42,7 +60,7 @@ export class ArtifactProcessor {
     return artifact;
   }
 
-  public static tryGetArtifactByBytecode(bytecode: string): ArtifactExtended {
+  public tryGetArtifactByBytecode(bytecode: string): ArtifactExtended {
     const artifact = ArtifactStorage.get(bytecodeHash(bytecode));
 
     if (!artifact) {
@@ -52,13 +70,13 @@ export class ArtifactProcessor {
     return artifact;
   }
 
-  public static tryGetContractName(bytecode: string): string {
+  public tryGetContractName(bytecode: string): string {
     const artifact = this.tryGetArtifactByBytecode(bytecode);
 
     return `${artifact.sourceName}:${artifact.contractName}`;
   }
 
-  private static _parseLibrariesOfArtifact(artifact: Artifact): NeededLibrary[] {
+  private _parseLibrariesOfArtifact(artifact: Artifact): NeededLibrary[] {
     const libraries = artifact.linkReferences;
 
     const neededLibraries = [];
@@ -77,7 +95,9 @@ export class ArtifactProcessor {
     return neededLibraries;
   }
 
-  private static _isNotDeployableArtifact(artifact: Artifact): boolean {
+  private _isNotDeployableArtifact(artifact: Artifact): boolean {
     return artifact.deployedBytecode === "0x";
   }
 }
+
+export const ArtifactProcessor = new BaseArtifactProcessor();