+++ test/fixture-projects/hardhat-project-minimal-truffle/hardhat.config.js
@@ -1,5 +1,0 @@
-require("../../../src");
-
-require("@nomiclabs/hardhat-truffle5");
-
-module.exports = {};