+++ test/fixture-projects/hardhat-project-typechain-truffle/hardhat.config.ts
@@ -1,0 +1,11 @@
+import { HardhatUserConfig } from "hardhat/config";
+
+import "@nomiclabs/hardhat-truffle5";
+
+import config from "../hardhat.config";
+
+const defaultConfig: HardhatUserConfig = {
+  ...config,
+};
+
+export default defaultConfig;