+++ test/integration/migration/pure-bytecode.test.ts
@@ -1,0 +1,30 @@
+import { HardhatRuntimeEnvironment } from "hardhat/types";
+
+import { runWithContinue, runWithoutContinue } from "./helper";
+
+import { useEnvironment } from "../../helpers";
+import { resetReporter } from "../../../src/tools/reporters/Reporter";
+import { resetTransactionRunner } from "../../../src/tools/runners/TransactionRunner";
+
+describe("pure-bytecode", () => {
+  useEnvironment("pure-bytecode");
+
+  let hre: HardhatRuntimeEnvironment;
+
+  beforeEach("setup", async function () {
+    hre = this.hre;
+
+    resetReporter();
+    resetTransactionRunner();
+  });
+
+  describe("simple migration flow", () => {
+    it("should run migration successfully", async function () {
+      await runWithoutContinue(hre, 1);
+    });
+
+    it("should recover migration successfully", async function () {
+      await runWithContinue(hre, 1);
+    });
+  });
+});